+++ angular-workspace/projects/component-library/src/lib/models/modal.models.ts
@@ -1,6 +1,6 @@
 import { TemplateRef, Type } from "@angular/core";
 
-export type ModalRole = "dialog" | "alert" | "alertDialog";
+export type ModalRole = "dialog" | "alert" | "alertdialog";
 
 export type AllowedModalContentTypes = Type<unknown> | TemplateRef<unknown>;
 
@@ -14,4 +14,5 @@ export interface ModalOptions {
   role?: ModalRole;
   showCloseButton?: boolean;
   initialFocus?: string;
+  returnFocus?: (e: HTMLElement | SVGElement) => HTMLElement | SVGElement | string | false;
 }